// apps/nativeapp/src/types/interface/slice.interface.ts
export interface PlantedSpecies {
  id?: string,
  guid: string,
  scientificName: string,
  aliases: string,
  count: number,
  image: string
}

export interface AdditionalDetail {
  key: string
  value: string
  access_type: string
}

export interface SampleTreeSlice {
  tree_id: string
  sample_tree_count: number
  boundary: Array<number[]>
  coordinates: Array<number[]>
  image_url: string
  current_species: PlantedSpecies,
  form_id: string
}